requirements.txt
Django==5.0.2
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-like==0.2.0
django-minio-backend==3.6.0
django-rest-framework==0.1.0
django-storage-celery-results==0.0.2b1
django-storages==1.14.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
drf-yasg==1.21.7
duplicity==0.8.21
email_validator==2.1.1
eradicate==2.3.0
executing==2.0.1
fonttools==4.48.1
frozenlist==1.4.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.42
